+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Show choices for array/combo options on hover
 - Build .deb files for Ubuntu 18.04/20.04/22.04 and Debian stable/testing/unstable in CI
 - Setup [APT Repo](https://github.com/JCWasmx86/swift-mesonlsp-apt-repo) for these distributions
+- Include identifiers in auto-completion
 # 2.0 (Apr 11 2023)
 - Only rebuild tree after loading subprojects if a subproject was found
 - Rebuild trees of folder based subprojects. This allows e.g. editing while preserving features like hover/document symbols